add PlayerController

+++ Tenebra/Assets/Scripts/IGenericSkillClass.cs
@@ -1,14 +1,14 @@
 
 public interface IGenericSkillClass 
 {
-    int CurrentLevel { get; }
-    double CurrentExp { get; set; }
-    double PreviousExpLevel { get;}
-    double NextExpLevel { get; }
+    long CurrentLevel { get; }
+    long CurrentExp { get; set; }
+    long PreviousExpLevel { get;}
+    long NextExpLevel { get; }
 
-    public double Formula(int level);
-    public void levelDown(double curtExp);
-    public void levelUp(double curtExp);
-    public void SetExp(int level);
+    public long Formula(long level);
+    public void levelDown(long curtExp);
+    public void levelUp(long curtExp);
+    public void SetExp(long level);
 
 }
